adiciona funções auxiliares para modelos Gemini e Ollama, validação da chave API e verificação do status do servidor Ollama; remove código não utilizado
