+++ gnnmodel/views.py
@@ -8,7 +8,7 @@ from django.http import HttpResponse, JsonResponse
 from django.shortcuts import render
 from django.views.decorators.http import require_http_methods
 
-from .forms import GoogleAPIKeyForm, InChIorSMILESareaInput, InChIorSMILESinput
+from .forms import InChIorSMILESareaInput
 from .models import ChatSession
 from .utils import (
     available_params,
@@ -20,7 +20,6 @@ from .utils import (
     process_mixture_post,
     process_pure_post,
 )
-from .utils_llm import resume_mol
 
 file_dir = osp.dirname(__file__)
 
@@ -92,33 +91,6 @@ def authorpage(request):
 def about(request):
     "handle request for about page"
     return render(request, "about.html")
-
-
-def description(request):
-    "handle request for molecule description"
-
-    html_output = ""
-
-    if request.method == "POST":
-        form = InChIorSMILESinput(request.POST)
-        google_api_key_form = GoogleAPIKeyForm(request.POST)
-        if google_api_key_form.is_valid() and form.is_valid():
-            smiles, inchi = form.cleaned_data["query"]
-            html_output = resume_mol(
-                inchi, smiles, google_api_key_form.cleaned_data["google_api_key"]
-            )
-    else:
-        form = InChIorSMILESinput()
-        google_api_key_form = GoogleAPIKeyForm()
-
-    return render(
-        request,
-        "description.html",
-        {
-            "output": html_output,
-            "google_api_key_form": google_api_key_form,
-        },
-    )
 
 
 def chat(request):
